project restructuring and added some styles for dark mode component
