+++ src/styles/dark-mode/dark-mode.css
@@ -1,0 +1,102 @@
+@import "../constatnts/constants.css";
+
+.new-tab__menu {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    position: absolute;
+    top: 0;
+    right: 0;
+    padding: 6px;
+}
+
+.new-tab__menu-icon-yandex,
+.new-tab__menu-icon-google,
+.new-tab__menu-icon-duckduckgo {
+    width: 34px;
+    height: 34px;
+    transform: rotate(90deg);
+    transition: var(--fast-transition);
+    cursor: pointer;
+}
+
+.new-tab__menu-icon-yandex {
+    fill: var(--ya-color)
+}
+
+.new-tab__menu-icon-google {
+    fill: var(--goo-color);
+}
+
+.new-tab__menu-icon-duckduckgo {
+    fill: var(--guck-color);
+}
+
+.new-tab__menu-box {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    visibility: hidden;
+    transition: var(--fast-transition);
+    opacity: 0;
+    font-size: 14px;
+}
+
+.new-tab__menu-box-corner {
+    position: relative;
+    width: 0;
+    right: 8px;
+    margin-top: 4px;
+    background-color: transparent;
+    border-right: 8px solid transparent;
+    border-left: 8px solid transparent;
+    border-bottom: 8px solid #fff;
+}
+
+.new-tab__menu-box-content {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 12px;
+    background-color: #fff;
+    border-radius: 4px;
+    box-shadow: 0 3px 6px 0 rgba(0,0,0,0.2);
+}
+
+.new-tab__menu-box-content h4 {
+    margin: 0;
+    font-size: 16px;
+}
+
+.new-tab__menu-box-content hr {
+    margin: 4px 0 8px;
+    width: 100%;
+}
+
+.new-tab__menu-box-content_dark-mode {
+    margin-bottom: 6px;
+    cursor: pointer;
+    width: 100%;
+    height: 24px;
+    border-radius: 4px;
+}
+
+.new-tab__menu-box-content_dark-switcher {
+    cursor: pointer;
+}
+
+
+.rotate {
+    transform: rotate(215deg);
+}
+
+.visible {
+    visibility: visible;
+    opacity: 1;
+}
+
+.disabled {
+    pointer-events: none;
+    opacity: 0.4;
+}
