+++ src/styles/new-tab/new-tab.css
@@ -11,6 +11,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    font-family: sans-serif;
 }
 
 .new-tab.dark {
