--- src/styles/new-tab/new-tab_toggle.css
@@ -1,200 +1,0 @@
-@import "../constatnts/constants.css";
-
-.new-tab__menu {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 6px;
-}
-
-.new-tab__menu-icon-yandex,
-.new-tab__menu-icon-google,
-.new-tab__menu-icon-duckduckgo {
-    width: 34px;
-    height: 34px;
-    transform: rotate(90deg);
-    transition: var(--fast-transition);
-    cursor: pointer;
-}
-
-.new-tab__menu-icon-yandex {
-    fill: var(--ya-color)
-}
-
-.new-tab__menu-icon-google {
-    fill: var(--goo-color);
-}
-
-.new-tab__menu-icon-duckduckgo {
-    fill: var(--guck-color);
-}
-
-.new-tab__menu-box {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    visibility: hidden;
-    transition: var(--fast-transition);
-    opacity: 0;
-}
-
-.new-tab__menu-box-corner {
-    position: relative;
-    width: 0;
-    right: 8px;
-    margin-top: 4px;
-    background-color: transparent;
-    border-right: 8px solid transparent;
-    border-left: 8px solid transparent;
-    border-bottom: 8px solid #fff;
-}
-
-.new-tab__menu-box-content {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 12px;
-    background-color: #fff;
-    border-radius: 4px;
-    box-shadow: 0 3px 6px 0 rgba(0,0,0,0.2);
-}
-
-.new-tab__menu-box-content h4 {
-    margin: 0;
-}
-
-.new-tab__menu-box-content hr {
-    margin: 4px 0;
-    width: 100%;
-}
-
-.new-tab__menu-box-content_dark-mode {
-    margin-bottom: 6px;
-    cursor: pointer;
-}
-
-.new-tab__menu-box-content_dark-switcher {
-    cursor: pointer;
-}
-
-.rotate {
-    transform: rotate(215deg);
-}
-
-.visible {
-    visibility: visible;
-    opacity: 1;
-}
-
-.disabled {
-    pointer-events: none;
-    opacity: 0.4;
-}
-
-.new-tab_toggle-btn {
-    position: fixed;
-    width: var(--width);
-    height: var(--height);
-    right: 10px;
-    top: 10px;
-    border-radius: 40px;
-}
-
-.new-tab_toggle-checkbox {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    cursor: pointer;
-    opacity: 0;
-    z-index: 2;
-}
-.new-tab_toggle-yandex,
-.new-tab_toggle-google,
-.new-tab_toggle-duckduckgo {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    overflow: hidden;
-    opacity: 1;
-    background-color: var(--light-color);
-    box-shadow: 0 2px 25px #d9d9d9;
-    border-radius: 40px;
-    transition: 0.2s ease background-color, 0.2s ease opacity;
-}
-
-.new-tab_toggle-yandex:before,
-.new-tab_toggle-google:before,
-.new-tab_toggle-duckduckgo:before,
-.new-tab_toggle-yandex:after,
-.new-tab_toggle-google:after,
-.new-tab_toggle-duckduckgo:after {
-    content: "";
-    position: absolute;
-    border-radius: 50%;
-    left: -6px;
-    top: calc(var(--height) * 0.11);
-    width: calc(var(--height) * 0.78);
-    height: calc(var(--height) * 0.78);
-    transition: 0.5s ease transform, 0.2s ease background-color;
-}
-
-.new-tab_toggle-yandex:before,
-.new-tab_toggle-google:before,
-.new-tab_toggle-duckduckgo:before {
-    background-color: var(--light-color);
-    transform: translate(-58px, 0px);
-    z-index: 1;
-}
-
-.new-tab_toggle-yandex:after {
-    transition: var(--fast-transition);
-    background-color: var(--ya-color);
-    transform: translate(8px, 0px);
-    z-index: 0;
-}
-
-.new-tab_toggle-google:after {
-    transition: var(--fast-transition);
-    background-color: var(--goo-color);
-    transform: translate(8px, 0px);
-    z-index: 0;
-}
-
-.new-tab_toggle-duckduckgo:after {
-    transition: var(--fast-transition);
-    background-color: var(--guck-color);
-    transform: translate(8px, 0px);
-    z-index: 0;
-}
-
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-yandex,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-google,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-duckduckgo {
-    background-color: var(--dark-color);
-}
-
-.new-tab_toggle-checkbox:active + .new-tab_toggle-yandex,
-.new-tab_toggle-checkbox:active + .new-tab_toggle-google,
-.new-tab_toggle-checkbox:active + .new-tab_toggle-duckduckgo {
-    opacity: 0.5;
-}
-
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-yandex:before,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-google:before,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-duckduckgo:before {
-    background-color: var(--dark-color);
-    transform: translate(33px, -3px);
-}
-
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-yandex:after,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-google:after,
-.new-tab_toggle-checkbox:checked + .new-tab_toggle-duckduckgo:after {
-    background-color: var(--light-color);
-    transform: translate(calc(var(--width) * 0.76), 0px);
-}
